2016-05-04 Gamdong project
there is 3 java code.

First MainActivity is really simple code, just you click the button , activity close.
And I want to watch button move, so start animation.
But this move is Just in this Activity.
Not move widget button.

Second MyWidget is update for widget class.
widget update time is 30min, but i want to update time is 1 second.
so I use handler for update time